commit history reset
